update old area overlap functions based on new datasets
