+++ geoprocessing.json
@@ -19,13 +19,13 @@
   ],
   "geoprocessingFunctions": [
     "src/functions/boundaryAreaOverlap.ts",
-    "src/functions/geomorphicOverlap.ts",
+    "src/functions/windFarmOverlap.ts",
     "src/functions/priorityAreaOverlap.ts",
     "src/functions/fishingEffortOverlap.ts",
     "src/functions/miningOverlap.ts",
     "src/functions/shippingOverlap.ts",
-    "src/functions/ecozonesPelagicOverlap.ts",
-    "src/functions/ecozonesBenthicOverlap.ts",
-    "src/functions/oilExplorationOverlap.ts"
+    "src/functions/powerLinesOverlap.ts",
+    "src/functions/bathyClassesOverlap.ts",
+    "src/functions/oilAreasOverlap.ts"
   ]
 }
