+++ project/datasources.json
@@ -146,28 +146,6 @@
     "src": "data/src/Brazil/Data_Products/south-brazilian-shelf-db/kernel_gillnet_preps.tif"
   },
   {
-    "datasourceId": "mining_active",
-    "geo_type": "vector",
-    "keyStats": {
-      "total": {
-        "total": {
-          "count": 543,
-          "sum": null,
-          "area": 1245836426.3570294
-        }
-      }
-    },
-    "formats": [
-      "fgb"
-    ],
-    "layerName": "mining_active_anm_clipped",
-    "classKeys": [],
-    "created": "2023-07-11T18:21:33.468Z",
-    "lastUpdated": "2023-08-08T17:27:38.305Z",
-    "src": "data/src/Brazil/Data_Products/south-brazilian-shelf-db/mining_active_anm_clipped.fgb",
-    "propertiesToKeep": []
-  },
-  {
     "datasourceId": "mining_potential",
     "geo_type": "vector",
     "keyStats": {
@@ -466,138 +444,6 @@
     "propertiesToKeep": [
       "intensity"
     ]
-  },
-  {
-    "datasourceId": "mining_active_central-east",
-    "geo_type": "vector",
-    "keyStats": {
-      "total": {
-        "total": {
-          "count": 164,
-          "sum": null,
-          "area": 539055700.3251766
-        }
-      }
-    },
-    "formats": [
-      "fgb"
-    ],
-    "layerName": "mining_active_central-east",
-    "classKeys": [],
-    "created": "2023-07-26T23:32:56.477Z",
-    "lastUpdated": "2023-08-08T17:27:44.007Z",
-    "src": "data/src/Brazil/Data_Products/region-test/mining_active_central-east.fgb",
-    "propertiesToKeep": []
-  },
-  {
-    "datasourceId": "mining_active_south",
-    "geo_type": "vector",
-    "keyStats": {
-      "total": {
-        "total": {
-          "count": 368,
-          "sum": null,
-          "area": 696540431.481384
-        }
-      }
-    },
-    "formats": [
-      "fgb"
-    ],
-    "layerName": "mining_active_south",
-    "classKeys": [],
-    "created": "2023-07-26T23:33:34.870Z",
-    "lastUpdated": "2023-08-08T17:27:44.248Z",
-    "src": "data/src/Brazil/Data_Products/region-test/mining_active_south.fgb",
-    "propertiesToKeep": []
-  },
-  {
-    "datasourceId": "mining_potential_central-east",
-    "geo_type": "vector",
-    "keyStats": {
-      "total": {
-        "total": {
-          "count": 28,
-          "sum": null,
-          "area": 61474305634.50131
-        }
-      }
-    },
-    "formats": [
-      "fgb"
-    ],
-    "layerName": "mining_potential_central-east",
-    "classKeys": [],
-    "created": "2023-07-26T23:34:34.387Z",
-    "lastUpdated": "2023-08-08T17:27:44.533Z",
-    "src": "data/src/Brazil/Data_Products/region-test/mining_potential_central-east.fgb",
-    "propertiesToKeep": []
-  },
-  {
-    "datasourceId": "mining_potential_north",
-    "geo_type": "vector",
-    "keyStats": {
-      "total": {
-        "total": {
-          "count": 15,
-          "sum": null,
-          "area": 65155946688.930725
-        }
-      }
-    },
-    "formats": [
-      "fgb"
-    ],
-    "layerName": "mining_potential_north",
-    "classKeys": [],
-    "created": "2023-07-26T23:35:12.032Z",
-    "lastUpdated": "2023-08-08T17:27:44.784Z",
-    "src": "data/src/Brazil/Data_Products/region-test/mining_potential_north.fgb",
-    "propertiesToKeep": []
-  },
-  {
-    "datasourceId": "mining_potential_northeast",
-    "geo_type": "vector",
-    "keyStats": {
-      "total": {
-        "total": {
-          "count": 86,
-          "sum": null,
-          "area": 98479921048.17107
-        }
-      }
-    },
-    "formats": [
-      "fgb"
-    ],
-    "layerName": "mining_potential_northeast",
-    "classKeys": [],
-    "created": "2023-07-26T23:35:44.679Z",
-    "lastUpdated": "2023-08-08T17:27:45.049Z",
-    "src": "data/src/Brazil/Data_Products/region-test/mining_potential_northeast.fgb",
-    "propertiesToKeep": []
-  },
-  {
-    "datasourceId": "mining_potential_south",
-    "geo_type": "vector",
-    "keyStats": {
-      "total": {
-        "total": {
-          "count": 35,
-          "sum": null,
-          "area": 42303438767.39601
-        }
-      }
-    },
-    "formats": [
-      "fgb"
-    ],
-    "layerName": "mining_potential_south",
-    "classKeys": [],
-    "created": "2023-07-26T23:36:42.040Z",
-    "lastUpdated": "2023-08-08T17:27:45.285Z",
-    "src": "data/src/Brazil/Data_Products/region-test/mining_potential_south.fgb",
-    "propertiesToKeep": []
   },
   {
     "datasourceId": "wind_farms",
