--- src/components/OilExplorationCard.tsx
@@ -1,184 +1,0 @@
-import React from "react";
-import {
-  Collapse,
-  ClassTable,
-  SketchClassTable,
-  ResultsCard,
-  useSketchProperties,
-  ToolbarCard,
-  LayerToggle,
-} from "@seasketch/geoprocessing/client-ui";
-import {
-  ReportResult,
-  toNullSketchArray,
-  flattenBySketchAllClass,
-  metricsWithSketchId,
-  toPercentMetric,
-  squareMeterToKilometer,
-  valueFormatter,
-} from "@seasketch/geoprocessing/client-core";
-
-import project from "../../project";
-import Translator from "./TranslatorAsync";
-import { Trans, useTranslation } from "react-i18next";
-
-const metricGroup = project.getMetricGroup("oilExplorationOverlap");
-const precalcMetrics = project.getPrecalcMetrics(
-  metricGroup,
-  "area",
-  metricGroup.classKey
-);
-
-const Number = new Intl.NumberFormat("en", { style: "decimal" });
-
-export const OilExplorationCard = () => {
-  const [{ isCollection }] = useSketchProperties();
-  const { t } = useTranslation();
-
-  const mapLabel = t("Map");
-  const typeLabel = t("Type");
-  const areaWithin = t("Area Within Plan");
-  const percAreaWithin = `% ${t("Area Within Plan")}`;
-  const sqKmLabel = t("km²");
-
-  return (
-    <>
-      <ResultsCard
-        title={t("Oil and Gas Exploration")}
-        functionName="oilExplorationOverlap"
-        useChildCard
-      >
-        {(data: ReportResult) => {
-          let singleMetrics = data.metrics.filter(
-            (m) => m.sketchId === data.sketch.properties.id
-          );
-
-          const finalMetrics = [
-            ...singleMetrics,
-            ...toPercentMetric(
-              singleMetrics,
-              precalcMetrics,
-              project.getMetricGroupPercId(metricGroup)
-            ),
-          ];
-
-          return (
-            <ToolbarCard
-              title={t("Oil and Gas Exploration")}
-              items={
-                <LayerToggle
-                  label={mapLabel}
-                  layerId={metricGroup.layerId}
-                  simple
-                />
-              }
-            >
-              <Translator>
-                <ClassTable
-                  rows={finalMetrics}
-                  metricGroup={metricGroup}
-                  columnConfig={[
-                    {
-                      columnLabel: typeLabel,
-                      type: "class",
-                      width: 30,
-                    },
-                    {
-                      columnLabel: areaWithin,
-                      type: "metricValue",
-                      metricId: metricGroup.metricId,
-                      valueFormatter: (val: string | number) =>
-                        Number.format(
-                          Math.round(
-                            squareMeterToKilometer(
-                              typeof val === "string" ? parseInt(val) : val
-                            )
-                          )
-                        ),
-                      valueLabel: sqKmLabel,
-                      width: 30,
-                    },
-                    {
-                      columnLabel: percAreaWithin,
-                      type: "metricChart",
-                      metricId: project.getMetricGroupPercId(metricGroup),
-                      valueFormatter: "percent",
-                      chartOptions: {
-                        showTitle: true,
-                        targetLabelPosition: "bottom",
-                        targetLabelStyle: "tight",
-                        barHeight: 11,
-                      },
-                      width: 30,
-                      targetValueFormatter: (
-                        value: number,
-                        row: number,
-                        numRows: number
-                      ) => {
-                        if (row === 0) {
-                          return (value: number) =>
-                            `${valueFormatter(value / 100, "percent0dig")} ${t(
-                              "Target"
-                            )}`;
-                        } else {
-                          return (value: number) =>
-                            `${valueFormatter(value / 100, "percent0dig")}`;
-                        }
-                      },
-                    },
-                    {
-                      columnLabel: mapLabel,
-                      type: "layerToggle",
-                      width: 10,
-                    },
-                  ]}
-                />
-              </Translator>
-
-              {isCollection && (
-                <Collapse title={t("Show by MPA")}>
-                  {genSketchTable(data)}
-                </Collapse>
-              )}
-
-              <Collapse title={t("Learn more")}>
-                <Trans i18nKey="Priority Areas Card - learn more">
-                  <p>
-                    {" "}
-                    This report summarizes the proportion of Active and
-                    potential mining areas within this plan.
-                  </p>
-                  <p>
-                    If MPA boundaries overlap with each other, the overlap is
-                    only counted once.
-                  </p>
-                </Trans>
-              </Collapse>
-            </ToolbarCard>
-          );
-        }}
-      </ResultsCard>
-    </>
-  );
-};
-
-const genSketchTable = (data: ReportResult) => {
-  // Build agg metric objects for each child sketch in collection with percValue for each class
-  const childSketches = toNullSketchArray(data.sketch);
-  const childSketchIds = childSketches.map((sk) => sk.properties.id);
-  const childSketchMetrics = toPercentMetric(
-    metricsWithSketchId(
-      data.metrics.filter((m) => m.metricId === metricGroup.metricId),
-      childSketchIds
-    ),
-    precalcMetrics
-  );
-  const sketchRows = flattenBySketchAllClass(
-    childSketchMetrics,
-    metricGroup.classes,
-    childSketches
-  );
-  return (
-    <SketchClassTable rows={sketchRows} metricGroup={metricGroup} formatPerc />
-  );
-};
