+++ src/components/RepresentationPage.tsx
@@ -1,16 +1,15 @@
 import React from "react";
 // import { GeomorphicCard } from "./GeomorphicCard";
 import { PriorityAreaCard } from "./PriorityAreaCard";
-import { EcozonesPelagicCard } from "./EcozonesPelagicCard";
-import { EcozonesBenthicCard } from "./EcozonesBenthicCard";
+import { EcozonesPelagicCard } from "./PowerLinesCard";
+import { BathyClassesCard } from "./BathyClassesCard";
 
 const ReportPage = () => {
   return (
     <>
       {/* <GeomorphicCard /> */}
       <PriorityAreaCard />
-      <EcozonesPelagicCard />
-      <EcozonesBenthicCard />
+      <BathyClassesCard />
     </>
   );
 };
