+++ src/functions/powerLinesOverlap.ts
@@ -1,0 +1,102 @@
+import {
+  Sketch,
+  Feature,
+  GeoprocessingHandler,
+  Metric,
+  Polygon,
+  ReportResult,
+  SketchCollection,
+  toNullSketch,
+  overlapFeatures,
+  rekeyMetrics,
+  getFlatGeobufFilename,
+  isInternalVectorDatasource,
+} from "@seasketch/geoprocessing";
+import { fgbFetchAll } from "@seasketch/geoprocessing/dataproviders";
+import bbox from "@turf/bbox";
+import project from "../../project";
+
+export async function powerLinesOverlap(
+  sketch: Sketch<Polygon> | SketchCollection<Polygon>
+): Promise<ReportResult> {
+  const box = sketch.bbox || bbox(sketch);
+  const metricGroup = project.getMetricGroup("powerLinesOverlap");
+
+  let cachedFeatures: Record<string, Feature<Polygon>[]> = {};
+
+  const polysByBoundary = (
+    await Promise.all(
+      metricGroup.classes.map(async (curClass) => {
+        if (!curClass.datasourceId) {
+          throw new Error(`Missing datasourceId ${curClass.classId}`);
+        }
+        const ds = project.getDatasourceById(curClass.datasourceId);
+        if (isInternalVectorDatasource(ds)) {
+          const url = `${project.dataBucketUrl()}${getFlatGeobufFilename(ds)}`;
+
+          // Fetch features overlapping with sketch, pull from cache if already fetched
+          const dsFeatures =
+            cachedFeatures[curClass.datasourceId] ||
+            (await fgbFetchAll<Feature<Polygon>>(url, box));
+          cachedFeatures[curClass.datasourceId] = dsFeatures;
+
+          // If this is a sub-class, filter by class name, exclude null geometry too
+          // ToDo: should do deeper match to classKey
+          const finalFeatures =
+            ds.classKeys.length > 0 &&
+              curClass.classId !== `${ds.datasourceId}_all`
+              ? dsFeatures.filter((feat) => {
+                return (
+                  feat.geometry &&
+                  feat.properties![ds.classKeys[0]] === curClass.classId
+                );
+              }, [])
+              : dsFeatures;
+
+          return finalFeatures;
+        }
+        return [];
+      })
+    )
+  ).reduce<Record<string, Feature<Polygon>[]>>((acc, polys, classIndex) => {
+    return {
+      ...acc,
+      [metricGroup.classes[classIndex].classId]: polys,
+    };
+  }, {});
+
+  const metrics: Metric[] = (
+    await Promise.all(
+      metricGroup.classes.map(async (curClass) => {
+        const overlapResult = await overlapFeatures(
+          metricGroup.metricId,
+          polysByBoundary[curClass.classId],
+          sketch
+        );
+        return overlapResult.map(
+          (metric): Metric => ({
+            ...metric,
+            classId: curClass.classId,
+          })
+        );
+      })
+    )
+  ).reduce(
+    // merge
+    (metricsSoFar, curClassMetrics) => [...metricsSoFar, ...curClassMetrics],
+    []
+  );
+
+  return {
+    metrics: rekeyMetrics(metrics),
+    sketch: toNullSketch(sketch, true),
+  };
+}
+
+export default new GeoprocessingHandler(powerLinesOverlap, {
+  title: "powerLinesOverlap",
+  description: "Calculate sketch overlap with priority area polygons",
+  executionMode: "async",
+  timeout: 600,
+  requiresProperties: [],
+});
