+++ src/functions/windFarmOverlapSmoke.test.ts
@@ -1,0 +1,23 @@
+/**
+ * @jest-environment node
+ * @group smoke
+ */
+import { windFarmOverlap } from "./windFarmOverlap";
+import {
+  getExamplePolygonSketchAll,
+  writeResultOutput,
+} from "@seasketch/geoprocessing/scripts/testing";
+
+describe("Basic smoke tests", () => {
+  test("handler function is present", () => {
+    expect(typeof windFarmOverlap).toBe("function");
+  });
+  test("windFarmOverlapSmoke - tests run against all examples", async () => {
+    const examples = await getExamplePolygonSketchAll();
+    for (const example of examples) {
+      const result = await windFarmOverlap(example);
+      expect(result).toBeTruthy();
+      writeResultOutput(result, "windFarmOverlap", example.properties.name);
+    }
+  }, 120000);
+});
